syslog tcp without tls fails
Configuring the settings for a remote host to use TCP without TLS
results in a faulty rsyslog native configuration. The latter includes
StreamDriverMode="1" StreamDriver="ossl" so that TLS is attempted, but
without the other required settings for this to succeed. The solution
is to only include these StreamDriver settings if there is TLS config
present by checking that the hash for these is not empty.

VRVDR-52943