Support syslog-enhanced configuration
VRVDR-53476

Support "system syslog-enhanced" configuration model.

The configuration provides a match-action set of rules.

Each rule can match on message attributes, including severity,
facility, or message contents.

Flags are supported, and can be set or cleared as a rule action
A rule can match on the status of a flag, set or cleared.

Rules can also apply rate limiting to messages, rate limitting can be
applied to messages when a specified flag is set, or alternatively to
all messages. This supports flexible rate limiting that can be applied
globally, per-target or message attribute.