Only one posix-match selector is allowed
There is no loss of expressiveness as other rules can always be added
with the other posix-match selectors for OR type behaviour (or using
alternations in the regex itself), and AND type behaviour achieved by
setting a flag in a rule and later a rule using multiple flag selectors.

We change the under lying implementation to AND posix-match
together if they were several present, to be consistent with behaviour
of of how the other selector behave (ie all selectors in a rule have to
match have to match).

In the future the max-elements restriction could be relaxed, and it
decide if OR or AND type behaviour is desired for combining them,
understanding their interaction with other selectors in the rule.

VRVDR-53918