Revert "Specify StateFile for imjournal module"
Specifying the StateFile results in numerous rsyslog errors shown in
the journal from the first time rsyslog is started after boot, such as
'Network is unreachable' and 'create UDP socket bound to device failed:
No such device'. It is problematic to delay this start, as it may be
valid if there is no vyatta syslog config. A solution is to revert the
change that resulted in including these startup errors in the journal.

This reverts commit 508dbb469807592ebb5f9d168145296a09f09a7c.

VRVDR-55014