Reduce the number of 'network is unreachable' errors in the journal
In the case of network unreachability, omfwd uses the rsyslog action
handling, which attempts the action 10 times, then suspends for 30s,
these parameters not being configurable. As each attempt results in 2
errors (and 2 action logs), this results in a total of 20 errors
per host. A solution to reduce the number of errors due to this burst
is to rate-limit rsyslog internal messages to 4 in a 1s interval.

VRVDR-55000